+++ examples/web-static.ps1
@@ -14,6 +14,7 @@ Server -Threads 2 {
 
     # listen on localhost:8085
     listen *:$Port http
+    logger terminal
 
     # set view engine to pode renderer
     engine pode