+++ tests/unit/Tools/CronParser.Tests.ps1
@@ -1,0 +1,218 @@
+$path = $MyInvocation.MyCommand.Path
+$src = (Split-Path -Parent -Path $path) -ireplace '[\\/]tests[\\/]unit[\\/]', '/src/'
+Get-ChildItem "$($src)/*.ps1" | Resolve-Path | ForEach-Object { . $_ }
+
+Describe 'Get-CronFields' {
+    It 'Returns valid cron fields' {
+        Get-CronFields | Should Be @(
+            'Minute',
+            'Hour',
+            'DayOfMonth',
+            'Month',
+            'DayOfWeek'
+        )
+    }
+}
+
+Describe 'Get-CronFieldConstraints' {
+    It 'Returns valid cron field constraints' {
+        $constraints = Get-CronFieldConstraints
+        $constraints | Should Not Be $null
+
+        $constraints.MinMax | Should Be @(
+            @(0, 59),
+            @(0, 23),
+            @(1, 31),
+            @(1, 12),
+            @(0, 6)
+        )
+
+        $constraints.DaysInMonths | Should Be @(
+            31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31
+        )
+
+        $constraints.Months | Should Be @(
+            'January', 'February', 'March', 'April', 'May', 'June', 'July',
+            'August', 'September', 'October', 'November', 'December'
+        )
+    }
+}
+
+Describe 'Get-CronPredefined' {
+    It 'Returns valid predefined values' {
+        $values = Get-CronPredefined
+        $values | Should Not Be $null
+
+        $values['@minutely'] | Should Be '* * * * *'
+        $values['@hourly'] | Should Be '0 * * * *'
+        $values['@daily'] | Should Be '0 0 * * *'
+        $values['@weekly'] | Should Be '0 0 * * 0'
+        $values['@monthly'] | Should Be '0 0 1 * *'
+        $values['@quaterly'] | Should Be '0 0 1 1,4,8,7,10'
+        $values['@yearly'] | Should Be '0 0 1 1 *'
+        $values['@annually'] | Should Be '0 0 1 1 *'
+        $values['@twice-hourly'] | Should Be '0,30 * * * *'
+        $values['@twice-daily'] | Should Be '0,12 0 * * *'
+        $values['@twice-weekly'] | Should Be '0 0 * * 0,4'
+        $values['@twice-monthly'] | Should Be '0 0 1,15 * *'
+        $values['@twice-yearly'] | Should Be '0 0 1 1,6 *'
+        $values['@twice-annually'] | Should Be '0 0 1 1,6 *'
+    }
+}
+
+Describe 'Get-CronFieldAliases' {
+    It 'Returns valid aliases' {
+        $aliases = Get-CronFieldAliases
+        $aliases | Should Not Be $null
+
+        $aliases.Month.Jan | Should Be 1
+        $aliases.Month.Feb | Should Be 2
+        $aliases.Month.Mar | Should Be 3
+        $aliases.Month.Apr | Should Be 4
+        $aliases.Month.May | Should Be 5
+        $aliases.Month.Jun | Should Be 6
+        $aliases.Month.Jul | Should Be 7
+        $aliases.Month.Aug | Should Be 8
+        $aliases.Month.Sep | Should Be 9
+        $aliases.Month.Oct | Should Be 10
+        $aliases.Month.Nov | Should Be 11
+        $aliases.Month.Dec | Should Be 12
+
+        $aliases.DayOfWeek.Sun | Should Be 0
+        $aliases.DayOfWeek.Mon | Should Be 1
+        $aliases.DayOfWeek.Tue | Should Be 2
+        $aliases.DayOfWeek.Wed | Should Be 3
+        $aliases.DayOfWeek.Thu | Should Be 4
+        $aliases.DayOfWeek.Fri | Should Be 5
+        $aliases.DayOfWeek.Sat | Should Be 6
+    }
+}
+
+Describe 'ConvertFrom-CronExpression' {
+    Context 'Invalid parameters supplied' {
+        It 'Throw null expression parameter error' {
+            { ConvertFrom-CronExpression -Expression $null } | Should Throw 'The argument is null or empty'
+        }
+
+        It 'Throw empty expression parameter error' {
+            { ConvertFrom-CronExpression -Expression ([string]::Empty) } | Should Throw 'The argument is null or empty'
+        }
+    }
+
+    Context 'Valid schedule parameters' {
+        It 'Throws error for too few number of cron atoms' {
+            { ConvertFrom-CronExpression -Expression '* * *' } | Should Throw 'Cron expression should only consist of 5 parts'
+        }
+
+        It 'Throws error for too many number of cron atoms' {
+            { ConvertFrom-CronExpression -Expression '* * * * * *' } | Should Throw 'Cron expression should only consist of 5 parts'
+        }
+
+        It 'Throws error for range atom with min>max' {
+            { ConvertFrom-CronExpression -Expression '* * 20-15 * *' } | Should Throw 'should not be greater than the max value'
+        }
+
+        It 'Throws error for range atom with invalid min' {
+            { ConvertFrom-CronExpression -Expression '* * 0-5 * *' } | Should Throw 'is invalid, should be greater than/equal to'
+        }
+
+        It 'Throws error for range atom with invalid max' {
+            { ConvertFrom-CronExpression -Expression '* * 1-32 * *' } | Should Throw 'is invalid, should be less than/equal to'
+        }
+
+        It 'Throws error for atom with invalid min' {
+            { ConvertFrom-CronExpression -Expression '* * 0 * *' } | Should Throw 'invalid, should be between'
+        }
+
+        It 'Throws error for atom with invalid max' {
+            { ConvertFrom-CronExpression -Expression '* * 32 * *' } | Should Throw 'invalid, should be between'
+        }
+
+        It 'Returns a valid cron object for predefined' {
+            $cron = ConvertFrom-CronExpression -Expression '@minutely'
+
+            $cron.Month.Values | Should Be $null
+            $cron.Month.Range.Min | Should Be 1
+            $cron.Month.Range.Max | Should Be 12
+            $cron.Month.Constraints[0] | Should Be 1
+            $cron.Month.Constraints[1] | Should Be 12
+            $cron.Month.Random | Should Be $false
+
+            $cron.DayOfWeek.Values | Should Be $null
+            $cron.DayOfWeek.Range.Min | Should Be 0
+            $cron.DayOfWeek.Range.Max | Should Be 6
+            $cron.DayOfWeek.Constraints[0] | Should Be 0
+            $cron.DayOfWeek.Constraints[1] | Should Be 6
+            $cron.DayOfWeek.Random | Should Be $false
+
+            $cron.Minute.Values | Should Be $null
+            $cron.Minute.Range.Min | Should Be 0
+            $cron.Minute.Range.Max | Should Be 59
+            $cron.Minute.Constraints[0] | Should Be 0
+            $cron.Minute.Constraints[1] | Should Be 59
+            $cron.Minute.Random | Should Be $false
+
+            $cron.Hour.Values | Should Be $null
+            $cron.Hour.Range.Min | Should Be 0
+            $cron.Hour.Range.Max | Should Be 23
+            $cron.Hour.Constraints[0] | Should Be 0
+            $cron.Hour.Constraints[1] | Should Be 23
+            $cron.Hour.Random | Should Be $false
+
+            $cron.Random | Should Be $false
+
+            $cron.DayOfMonth.Values | Should Be $null
+            $cron.DayOfMonth.Range.Min | Should Be 1
+            $cron.DayOfMonth.Range.Max | Should Be 31
+            $cron.DayOfMonth.Constraints[0] | Should Be 1
+            $cron.DayOfMonth.Constraints[1] | Should Be 31
+            $cron.DayOfMonth.Random | Should Be $false
+        }
+
+        It 'Returns a valid cron object for expression' {
+            $cron = ConvertFrom-CronExpression -Expression '0/10 * * * 2'
+
+            $cron.Month.Values | Should Be $null
+            $cron.Month.Range.Min | Should Be 1
+            $cron.Month.Range.Max | Should Be 12
+            $cron.Month.Constraints[0] | Should Be 1
+            $cron.Month.Constraints[1] | Should Be 12
+            $cron.Month.Random | Should Be $false
+
+            $cron.DayOfWeek.Values | Should Be 2
+            $cron.DayOfWeek.Range.Min | Should Be $null
+            $cron.DayOfWeek.Range.Max | Should Be $null
+            $cron.DayOfWeek.Constraints[0] | Should Be 0
+            $cron.DayOfWeek.Constraints[1] | Should Be 6
+            $cron.DayOfWeek.Random | Should Be $false
+
+            $cron.Minute.Values | Should Be @(0, 10, 20, 30, 40, 50)
+            $cron.Minute.Range.Min | Should Be $null
+            $cron.Minute.Range.Max | Should Be $null
+            $cron.Minute.Constraints[0] | Should Be 0
+            $cron.Minute.Constraints[1] | Should Be 59
+            $cron.Minute.Random | Should Be $false
+
+            $cron.Hour.Values | Should Be $null
+            $cron.Hour.Range.Min | Should Be 0
+            $cron.Hour.Range.Max | Should Be 23
+            $cron.Hour.Constraints[0] | Should Be 0
+            $cron.Hour.Constraints[1] | Should Be 23
+            $cron.Hour.Random | Should Be $false
+
+            $cron.Random | Should Be $false
+
+            $cron.DayOfMonth.Values | Should Be $null
+            $cron.DayOfMonth.Range.Min | Should Be 1
+            $cron.DayOfMonth.Range.Max | Should Be 31
+            $cron.DayOfMonth.Constraints[0] | Should Be 1
+            $cron.DayOfMonth.Constraints[1] | Should Be 31
+            $cron.DayOfMonth.Random | Should Be $false
+        }
+    }
+}
+
+
+
+
+